+++ OfficeCrawler/Areas/Area.cs
@@ -1,0 +1,9 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace OfficeCrawler.Areas {
+    public class Area {
+
+    }
+}
