+++ OfficeCrawler/Interfaces/IEntity.cs
@@ -1,0 +1,8 @@
+using Microsoft.Xna.Framework;
+
+namespace OfficeCrawler {
+    interface IEntity {
+        void Update(GameTime gameTime);
+        void Draw(GameTime gameTime);
+    }
+}
