+++ OfficeCrawler/Interfaces/IRenderable.cs
@@ -1,0 +1,8 @@
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
+
+namespace OfficeCrawler {
+    interface IRenderable {
+        public abstract void Draw(SpriteBatch spriteBatch, GameTime gameTime);
+    }
+}
