+++ OfficeCrawler/Physics/Collider.cs
@@ -1,0 +1,9 @@
+using Microsoft.Xna.Framework;
+
+namespace OfficeCrawler {
+
+    public class Collider {
+
+        public Rectangle BoundingBox { get; }
+    }
+}
